Eliminated DbConnectionPool.NetCoreApp.cs
